+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `import-posts` command for importing archived posts.
+- Fetch activities from origin if signer doesn't match actor.
 
 ### Changed